fixed alignment issues
When I tried to use flex to align title and list together,
only the title centered and the list was pushed onto a new line.
I put the list in a separate container and that worked so I tried again
after commenting the code that did work, by having the title and list in
the same div/container. Worked this time. Not sure why? Maybe I had
closing tag in the wrong place? FLummoxed by something so simple :'C